move internal/painter.svgToImage to internal/svg.ToImage
